move RectTransform specific extensions to own class
